+++ lib/NotepadType.dart
@@ -35,18 +35,30 @@ class NotepadType {
     mtu = await _bleType.requestMtu(expectedMtu) - GATT_HEADER_LENGTH;
   }
 
+  void configConnectionPriority(BleConnectionPriority bleConnectionPriority) {
+    _bleType.requestConnectionPriority(bleConnectionPriority);
+  }
+
   void sendRequestAsync(String messageHead, Tuple2<String, String> serviceCharacteristic, Uint8List request) async {
     _bleType.writeValue(serviceCharacteristic, request);
     print('on${messageHead}Send: ${hex.encode(request)}');
   }
 
   Stream<Uint8List> receiveValue(Tuple2<String, String> serviceCharacteristic) =>
-      _bleType.inputValueStream.where((cv) => cv.item1 == serviceCharacteristic.item2).map((cv) => cv.item2);
+      _bleType.inputValueStream.where((cv) {
+        return cv.item1 == serviceCharacteristic.item2 || '0000${cv.item1}-$GSS_SUFFIX' == serviceCharacteristic.item2;
+      }).map((cv) => cv.item2);
 
   Future<Uint8List> receiveResponseAsync(String messageHead, Tuple2<String, String> serviceCharacteristic, Predicate intercept) async {
     var response = await receiveValue(serviceCharacteristic).firstWhere(intercept);
     print('on${messageHead}Receive: ${hex.encode(response)}');
     return response;
+  }
+
+  Future<T> fetchProperty<T>(Tuple2<String, String> serviceCharacteristic, Handle<T> handle) async {
+    _bleType.readValue(serviceCharacteristic);
+    var value = await receiveResponseAsync('Property', serviceCharacteristic, (value) => true);
+    return handle(value);
   }
 
   Future<T> executeCommand<T>(NotepadCommand<T> command) async {